* Moved Soprano/QRDF to kdesupport * Moved libKnepomuk, libKonto, and libKMetaData into kdelibs   (the GUI elements still need to be sorted out.) * Moved the KNepomuk coreservices and the service registry into   kdebase/runtime.
It was agreed that this move makes sense even though KDE should be frozen
already since it is an important technology.

All of the new modules are optional which means that if Soprano is not installed
none of the KNepomuk parts will be. Still missing is a cmake message
stating this.

CCMAIL: nepomuk-kde@semanticdesktop.org

svn path=/trunk/kdesupport/soprano/; revision=659840