We now have a fully functional Sesame2 backend. The only thing Sesame2 does not support are SPARQL ask queries (strange, seems easy to me) And also in the test the RDQL query fails which is not a surprise. This is where one entry in the TODO comes to play: we have no way of knowing which model supports which query languages.
To install and use it is most likely necessary to set JAVA_HOME to an installation of Java 1.6.0.
Sesame2 needs java 1.5 but I only tested 1.6.

I simply bundeled a current version of Sesame2 (and dependancies) with soprano. It seemed to the the
easiest way to handle the stupid java system.

Well, I hope that was the last I saw of JNI for a while. Hopefully I can now test my InferencingEngine for real.

BTW: We should improve our feature-system and have some way to enable and disable features on backends or models.
     A good example is inferencing: if the model supports it we can choose to use that instead of our own 
     InferencingFilter.

CCMAIL: daniele.galdi@gmail.com

svn path=/branches/work/soprano2/; revision=700046