Do not depend on Environment variables on Windows
To deploy soprano on a Windows system one can not rely
on CMAKE_INSTALL_PREFIX. Because it is set during compile time
and will not be the same on a "Users" System who installs a soprano binary
package.
Using environment variables for this should be optional and not mandatory
and they are also not supported to be set by the KDE-Windows installer.

This Patch will use the location of the soprano executable or any
executable linking this code to look up the relative location of the
sopranodirs.

To avoid interference during the build and on a buildsystem the
old behavior of using the INSTALL_PREFIX is still left active.

Review: 101119
Reviewed by: S. Trueg