Revert "ErrorCache: Use QThreadStorage instead of a QHash with a mutex"
This reverts commit f4d2b9898ff958db1c69ac55fb21677bb6075e24.

QThreadStorage are only meant to be used for static variables. This is
not their intended use case. We get warning every time a QThreadStorage
is deleted before its thread is deleted. The ErrorCache is used heavily
in Soprano for small objects such as iterators, and thus we get these
warnings a lot.