Virtuoso Backend: Make empty graph support optional
Virtuoso doesn't support empty graphs, so we use a fake graph which acts
as the default empty graph.

This approach, while useful, does have a performance impact, as the
graph needs to be regularly checked against this default graph.

Therefore, we add an option to disable this feature. By default, this
feature is enabled in order to preserve compatibility.

REVIEW: 106520