Combine Server and Client streams into a new DataStream class
Instead of both the server and the client implementing their own
streaming operators. We combine all the streaming operators into a new
public class called Soprano::DataStream.

The Server and Client create their own respective class -
ServerDataStream and SocketStream respectively. The just have to
implement a read and write function.

This serves to eliminate code duplication, and allows others to use the
DataStream class, if they need it.

REVIEW: 106568