LocalSocketClient: Never delete the ClientConnection
The ClientConnection is not reference counted, and is passed to the the
model and open iterators. Deleting it when has open iterators therefore
leads to crashes.

Instead, the ClientConnection is always persistent. Intead we provide
functions to allow users of the ClientConnection to connect and
disconnect.

Also removed the concept of newConnection() from the ClientConnection.
It no longer spawns a new socket for each thread, so there is no point
of it existing. This new design is a lot simpler.

REVIEW: 107531