VirtuosoBackend: Change the default virtuoso port from 1111 -> 1113
This is being done as per debian's request. Debian on installing the
virtuoso package, launches virtuoso on the default port (1111) and
attempts to change the pasword. If Nepomuk is running, it inteferes with
their launch and causes problems.

The *apparently* cannot launch on anything but the default port, and
would like us to start on a different port.

CCMAIL: panfaust@gmail.com