Server: Add a stop method which closes all servers
Also had to change the internal signal emitted by the ServerConnection.
They used to emit finished() which would then result in us removing them
from the server list of connections. This would be done using the
'QObject::sender()'. However, since the stop signal deletes the
connection the qobject_cast fails and gives 0. This result in the
connection not being removed from the list. Also, when stop() is called
the next time we try to delete it again.

Instead we now connect to the destroyed(QObject*) signal, and avoid
using qobject_cast.