Virtuoso Backend: Optimize ODBC getCharData
Only use 1 SQLFetchData command in most of the cases.

Callgrind stats show that 67.5% of the time in this function is spent in
the first SQLFetchData, and an additional 27% in the second SQLGetData.
We can avoid some of this extra cost, by only calling the function
once.

Also, doing a simple fetch of 50000 quads -
Before Patch - 5.75 seconds
After Patch  - 4.70 seconds

REVIEW: 106467