VirtuosoBackend: Avoid calling SQLGetDescField unnecessarily
When fetching a literal node, it can either have a type or have a
language. We should not attempt to fetch both the langauge and type and
then check which one succeded.

Instead we first fetch the type, and only if that fails do we fetch the
lanaguge. This way we avoid one extra SQLGetDescField call when it is
not required.